Added Graviton and Radeon support

+++ templates/userdata.tpl
@@ -16,7 +16,12 @@ aws ec2 describe-tags --region $REGION --filter "Name=resource-id,Values=$INSTAN
 export ManifestPath=$ManifestPath
 echo $ManifestPath > /root/manifestpath
 
-yum install -y jq
+if [[ -f $(which yum) ]]; then
+	yum install -y jq
+else
+	apt-get update
+	apt-get install -y jq wget p7zip-full maskprocessor
+fi
 
 export BUCKET=${dictionaryBucket}
 export BUCKETREGION=${userdataRegion}
@@ -26,19 +31,32 @@ echo "Using dictionary bucket $BUCKET";
 mkdir /potfiles
 
 # format & mount /dev/xvdb
-mkfs.ext4 /dev/xvdb
-mkdir /xvdb
-mount /dev/xvdb /xvdb/
-mkdir /xvdb/npk-wordlist
-ln -s /xvdb/npk-wordlist /root/npk-wordlist
+if [[ -f /dev/xvdb ]]; then
+	mkfs.ext4 /dev/xvdb
+	mkdir /xvdb
+	mount /dev/xvdb /xvdb/
+	mkdir /xvdb/npk-wordlist
+	ln -s /xvdb/npk-wordlist /root/npk-wordlist
+else
+	mkfs.ext4 /dev/nvme1n1
+	mkdir /nvme1n1
+	mount /dev/nvme1n1 /nvme1n1/
+	mkdir /nvme1n1/npk-wordlist
+	ln -s /nvme1n1/npk-wordlist /root/npk-wordlist
+fi;
 
-aws s3 cp s3://$BUCKET/components-v3/epel.rpm .
-aws s3 cp s3://$BUCKET/components-v3/hashcat.7z .
-aws s3 cp s3://$BUCKET/components-v3/maskprocessor.7z .
 aws s3 cp s3://$BUCKET/components-v3/compute-node.7z .
 aws s3 cp s3://$USERDATA/$ManifestPath/manifest.json .
-rpm -Uvh epel.rpm
-yum install -y p7zip p7zip-plugins
+
+if [[ -f $(which yum) ]]; then
+	aws s3 cp s3://$BUCKET/components-v3/epel.rpm .
+	rpm -Uvh epel.rpm
+	yum install -y p7zip p7zip-plugins
+fi
+
+if [[ `lspci | grep AMD | wc -l` -gt 0 ]]; then
+	yum install -y opencl-amdgpu-pro
+fi
 
 # Install nvm
 curl -o- https://raw.githubusercontent.com/creationix/nvm/v0.33.2/install.sh | /bin/bash
@@ -69,7 +87,6 @@ jq -r '.dictionaryFile' manifest.json | grep \.gz$ | xargs -L1 -I'{}' gunzip ./n
 
 jq -r '.rulesFiles[]' manifest.json | grep \.7z$ | xargs -L1 -I'{}' 7z x ./npk-{} -o./npk-rules/
 jq -r '.rulesFiles[]' manifest.json | grep \.gz$ | xargs -L1 -I'{}' gunzip ./npk-{}
-
 
 ls -alh ./npk-rules/
 
@@ -103,10 +120,21 @@ aws ec2 describe-spot-fleet-instances --region $REGION --spot-fleet-request-id $
 export INSTANCECOUNT=$(cat fleet_instances | wc -l)
 export INSTANCENUMBER=$(cat fleet_instances | grep -nr $INSTANCEID - | cut -d':' -f1)
 
-7z x hashcat.7z
-7z x maskprocessor.7z
-mv hashcat-*/ hashcat
-mv maskprocessor-*/ maskprocessor
+if [[ -f $(which yum) ]]; then
+	aws s3 cp s3://$BUCKET/components-v3/hashcat.7z .
+	aws s3 cp s3://$BUCKET/components-v3/maskprocessor.7z .
+
+	7z x hashcat.7z
+	7z x maskprocessor.7z
+	mv hashcat-*/ hashcat
+	mv maskprocessor-*/ maskprocessor
+else
+	aws s3 cp s3://$BUCKET/components-v3/hashcat.arm64.7z .
+	7z x hashcat.arm64.7z
+	mv /root/hashcat/hashcat /root/hashcat/hashcat.bin
+	mkdir /root/maskprocessor
+	ln -s $(which mp64) /root/maskprocessor/mp64.bin
+fi
 
 7z x compute-node.7z
